Simplify MaterialButton.setTextAndIcon() logic
Also fix BlendModeColorFilter only being available on API 29+

(cherry picked from commit 4253af24fd1b7211e0421563f124e305bdda45a9)
